feat: 別コンポーネントの切り分け
エレメントの再生成を防ぐため
